fix(demo): run kanban demo in an infinite loop
This commit addresses an issue where the producer in the Kanban demo
would halt after generating a certain number of jobs. Earlier, even if
the producer was restarted, it wouldn’t resume the job generation due to
the presence of a '.done' marker file.
The marker file mechanism was therefore replaced with a simple polling
approach. Now restarting the producer works and will result in new jobs
being generated.

Signed-off-by: Michael Adler <michael.adler@siemens.com>