chore(deps): raise minimum supported Go version to 1.19
This commit updates the minimum supported version of Go to
1.19. The primary reason for this upgrade is that Go 1.18
lacks support for atomic.Bool, which is useful for the
implementation of job status subscriptions. Last but not
least, the major version of Go is now the same as used in Debian
Bookworm, the current Debian stable version (and there's no need to
support older versions than Debian).

Signed-off-by: Michael Adler <michael.adler@siemens.com>