ci: adjust leniency in hitting the patch target
Due to Go's inherently verbose error checking it is sometimes not
practical to test *all* possible error scenarios, e.g. there is no need
to test wrapper code such as

```go
if err != nil {
    return nil, err
}
```

With this commit, we allow some leniency regarding the *patch* coverage.
We still ensure that the *overall* coverage does not decrease though.

Signed-off-by: Michael Adler <michael.adler@siemens.com>