fix: include workflow definition in response when creating jobs
When querying one or multiple jobs, the response always includes the
workflow definition. However, this was not the case when creating a new
job; the response omitted the workflow definition. This inconsistency
not only diverged from our established response behavior but also from
the examples generated in our Swagger documentation. This change ensures
that the response for newly created jobs includes the workflow
definition, aligning it with our existing standards and documentation.

Reported-by: Robin Warth <robin.warth@siemens.com>
Signed-off-by: Michael Adler <michael.adler@siemens.com>