ci: exclude loadtest and viewer from code coverage
This commit excludes both the loadtest and viewer modules from coverage
data collected by codecov. A recent change in Codecov's behavior has
unexpectedly included these modules in coverage tracking. While there
may be an opportunity to implement tests for wfx-viewer in the
future, the primary objective of this commit is to revert the code
coverage metrics to their original state.

Signed-off-by: Michael Adler <michael.adler@siemens.com>