ci: remove +git version suffix from released *.deb packages
This commit eliminates the +git version suffix from the versioning of
our released Debian packages. The removal was done to ensure consistency
and clarity in our package versioning scheme, as the +git suffix could
cause confusion regarding the package's stability and readiness for
production use. The inspection of the packages via `dpkg -I` no longer
shows the +git version suffix.

Signed-off-by: Michael Adler <michael.adler@siemens.com>