ci: lower codecov severity to informational
Instead of failing the CI pipeline when code coverage slightly
decreases, this change sets the codecov severity to informational. This
allows the reviewer to assess the coverage changes and make the decision
on whether to accept the changes, rather than automatically failing the
build. This provides more flexibility and context for the coverage
changes.

Signed-off-by: Michael Adler <michael.adler@siemens.com>