fix: baseMWs are not called
The base middlewares (baseMWs) were not being invoked due to an
incorrect use of the copy() function, which only copies elements up to
the length of the destination slice. Since the destination slice was
initialized with zero length, no elements were copied. This has been
fixed by appending each middleware function individually to the
combinedMWs slice.

Fixes: f9825d67c1652e2b36cab526d7aa7d9bcd85ce85
Signed-off-by: Michael Adler <michael.adler@siemens.com>