refactor!: Remove Makefiles in favor of `go build`
Switch to using `go build` directly to embrace the Go way of building.
This change eliminates the need to maintain both the Makefile and the
goreleaser yaml. You can still build everything at once using the
provided justfile with `just build`.

Additionally, remove the go.mod files in subdirectories to avoid manual
updates when the main go.mod is updated. As a result, the linter now
also checks the contrib files, necessitating a few trivial fixes.

Signed-off-by: Michael Adler <michael.adler@siemens.com>