Start service only if systemd is running
Enable service if systemctl is available and start service only
if systemd is used as a running init process.
This split is required, since creating the link to the unit file
(enabling the service) is also desired to be created when creating
a rootfs-image "offline" (without systemd running at that time).