chore: proactively remove disconnected subscribers from event notifier
Previously, when a client disconnected, the event notifier was not
immediately informed. This led to repeated failed notification attempts
for disconnected subscribers, wasting time and memory.

With this change, we now notify the event notifier as soon as a client
disconnects, allowing it to promptly remove the subscriber and avoid
unnecessary notification attempts.

Signed-off-by: Michael Adler <michael.adler@siemens.com>